refactor: move consent-manager to one file and move it from theme to project level
